Add `OPTIONS` attribute to swift.Store function
Swift uses a function to wrap the final store instance and choose
between a SingleTenant/MultiTenant store. This patch adds the `OPTIONS`
attribute, found in all other store classes, in order to be able to
register the store options without actually creating an instance of the
Store - just like in every other store.

Change-Id: I581cdf5e8e5c112e129c2f0d62ad2a9caf19383b