VMware store: Re-use api session token
The store object is currently instanciated for each API call. The
constructor of the store object triggers the configuration of the
store. The session token to communicate with vCenter is generated
in this configuration step which can take some time depending on
the network.

This patch prevents the creation of the session throughout the API
calls.

Change-Id: Ic9520fa856264fda1ef3cb05f1395a8002192499
Closes-Bug: #1350010