Define a new parameter to pass CA cert file
This change adds a new parameter for the swift store driver that allows
to speficy the name of the CA cert file to use in the SSL connections for
verifying certificates. This parameter is passed to the swiftclient in
the creation of the connection. The parameter is called "swift_store_cacert".
This change corresponds to change
I4cbfae3c1ac84d6c85875d34a58dd2a87ae85d6f in glance.

Change-Id: I5b356170ec82d033204e22f79c862201400a0a31
Closes-bug: 1375857
DocImpact