Throw NotFound exception when template is gone
When using vsphere backend store, if the template is deleted in
vCenter datastore by some reason, it fails to delete the image
from glance due to the generic Exception.

The bug is fixed by catching oslo.vmware.FileNotFoundException,
and throw NotFound exception. So that the behavior of deleting
image will be identical with file system backend.

Change-Id: If22f71feadec025eaf9a38c4cb15eb7f669820f8
Closes-Bug: 1418396