Support for deleting images stored as SLO in Swift
When an image is deleted, the image data is not being deleted
completely if it is stored as static large object(SLO). Only the
manifest is deleted leaving behind all the segments. This leads to
stale data in swift cluster.

This change attempts to delete the SLO in the following way:
 1. Identify a SLO with 'x-static-large-object' header on an object
 2. Delete the SLO, both manifest and the segments, using the
 query string 'multipart-manifest=delete' while deleting the object

Closes-Bug: 1425617

Change-Id: Id2cd3b63ca88d66c5e7435b528ba75fa3c2cbb1c