Implement get, add and delete for cinder store
This implements get, add, delete methods for cinder storage
driver to enable users to upload/download the images to/from
the volumes. This also will be useful for users who want to
utilize storage features such as copy-on-write cloning
for creating a new volume from an image volume efficiently.

Change-Id: Ie228d1d95365e81881691af4cb8b170335a73868
Implements: blueprint cinder-store-upload-download