Fix RBD delete image on creation failure
When an exception rises on RBD store while adding/creating an image, and
the image has already been created, this new image is not properly
deleted as it should be.

Fault lies in incorrect call to Store._delete_image method, as it is
called with the wrong arguments.

Change-Id: Ib74f7eafbb04ab893039b480a7d3eaa15e7d59d1
Closes-Bug: #1449639
(cherry picked from commit 751f9ae444d25f9beaffee2796095102f884fe8d)