Add explicit dependencies for store dependencies
This adds explicit dependencies for the stores in the extras section.
Users can now install glance_store and specify which backends to install
dependencies. For example:

    $ pip install glance_store[swift]

should install glance_store and the dependences to use the swift store
(currently python-swiftclient and httplib2).

This has a number of benefits, primarily that if the dependencies for a
particular store change, the installation process should remain the
same.

Change-Id: I94d75e3d52c9e3e9f7a741a079d25b1bdff29b27