Sheepdog: Change storelocation format
This patch changes storelocation format from 'sheepdog://image' to
'sheepdog://addr:port:image'. It add ip and port to the location.
This offers the advantages of:

1.The ip and port represents a sheepdog cluster which contains the
image. The format like 'sheepdog://addr:port:image' can describe image
location accurately.

2.When clone an image in cinder, it need ip and port to decide whether
the image is in the same sheepdog cluster with cinder-backend, this is
very important for us to determine if the image is cloneable.

This patch also change a long data type to int to compatible with
python version 3.x.

Change-Id: Id79487e05e3af2f7b1287476ddbb174ad9de78c8