Fix default mutables for set_acls
read_tenants, write_tenants parameters in set_acls method of
glance_store/driver.Store have empty lists as default values.
It is not recommended approach cause it can lead to some neat
issues.

Moreover, Swift store class rewrites this method and uses None as
default values (which is right). So we need to follow this
approach in base Store class.

Change-Id: Ifa0f0c286dbb822b2d34d03978655d720bb109d4