Change Swift zero-size chunk behaviour
There is a difference in how Swift and radosgw behave when a zero
length write is aborted. Swift returns 499 and doesn't create an object,
but radosgw appears to creates a zero size object. Because of this, new
behaviour which was introduced to prevent zero-size chunks being written
worked for Swift, but appears not to work for the radosgw case.

This patch changes how we detect a zero size chunk so that no request
is made to the backend. We no longer rely on the Swift specific behaviour.

Change-Id: I1a48e2eceaa6b07b8e5bbfd42fef6fd3d91f5834
Closes-bug: 1537721