Remove S3 driver
As per the deprecation indicated in change
I139c36c2d591a55f0f91ea32efa963f80c28e444 , we are on track to removing
the S3 driver from glance_store in Newton.

The process of maintainers for individual drives was initiated in [1].
However, S3 doesn't have a maintainer and has been deprecated in Mitaka.

This commit intends to remove the S3 driver completely from the
glance_store source tree. If required, we recommend that this driver be
maintained outside of the glance_store source tree until you or someone
you know intend to support the code for long term. In the later case, we
hope that you will welcome our spec process to help plan the project
priorities.

DocImpact
UpgradeImpact

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-December/081966.html

Change-Id: I032b0fc16400cbd2112687d38e010128be699221