Fix help text of swift_store_config_file
While attempting to improve help text for configuration option
"swift_store_config_file", None is incorrectly listed as a
possible value. None is incorrect because a configuration option
can either be set or not. It can't be set to None. When no value
is set for a configuration option, it is inherently parsed as None
by oslo.config. Hence, there is no need to specify None as a
possible value.

Also, the configuration option swift_store_config_file is marked as
secret. Although the file pointed by this path can contain
credentials, the path itself need not be marked secret. This in fact
may make debugging harder if the file path is not logged clearly.

Change-Id: If4fa246d642098e96093cb6be92b8cd9f4350268
Partial-Bug: #1570946