Add cinder_volume_type to cinder store configuration
Some cinder backend can have several volume types to optimize
storage usage.
Adding this option allows to choose specific volume type in cinder
that can be optimized for images.
Now default_volume_type from cinder configuration is used for
volume creation but this solution is not so suitable for storing
images in cinder.

Closes-Bug: #1616556
Change-Id: I23a8cff7f7aa77acb418f5ad7dd785a893de8608