Reason to return sorted list of drivers for opts
This commit adds a note specifying the reason for returning the sorted
list of opts, this was missed in change
I3146b8597a5d89da49b84d6653edacc3067c2c71. The addition is being done in
a private method and we don't have unit tests for those. Having a note
in place will ensure no regressions.

Co-Authored-By: Nikhil Komawar <nik.komawar@gmail.com>
Co-Authored-By: Hemanth Makkapati <hemanth.makkapati@rackspace.com>

Change-Id: I4d9563166f8d81098d4d3d57ced7d7a63a6c01df
Related-Bug: 1619487