Fix option load for swift/vmware
There is a small problem in _option_get function, it has problem
dealing with boolean parameters with False value, such as 0, False.
It will rasie exception for missing options.

Fix the issue by changing the condition from "not result" to
"result is None" which is true indication for missing options.

Change-Id: I313aa9f8d583c9cf44b052685b0f26824d2b6ade