Add S3 store support to glance_store
Revive the S3 driver that was supported up to Mitaka and add multiple
store support to correspond with the features added to glance_store from
Newton to U cycle. In addition, the use of the S3 SDK used by the Mitaka
S3 driver has been deprecated, so this implementation will be changed
to use the new version.

Change-Id: I203134837319080ead96da69048baf90086d2117
Implements: blueprint re-support-s3-driver