Refactor methods in cinder store
Some of the methods in cinder store are only used by the Store class
methods and not used outside.
It makes sense to bound them to the class.
This refactoring will also help in handling the None imports inside
the class configure methods which will be done in a followup.

This patch also updates cinderclient version from v2(deprecated) to
v3(current).

Change-Id: I576b787012e73ebaf3274cfd8c31f79894284891