Don't allow image creation with encrypted nfs volumes
With change[1], we will enabled encrypted volume support for cinder
nfs driver.
Before enabling that, we require handling in cinder consuming projects
(like glance_store) to not have unexpected behavior.
This patch blocks image creation if the destination volume (in which
the image will be written) is an encrypted nfs volume.

[1] https://review.opendev.org/#/c/597148/

Closes-Bug: #1884482

Change-Id: I3b4c6782e561f22a7118624e26bf056cd6405a30