Fix mock import in unit tests
We allowed the use of the third party mock lib, even though we no longer
declare it as a test-requirement. That should have failed, but there
must still be a dependency that pulls it in for us. But unit tests will
start failing once more dependencies move away from it, so we need to
clean that up.

Change-Id: I79d647e8c8a060eacfa5cbee39a99344adc0ae94
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>