Support Cinder multiple stores
This patch adds support for configuring multiple cinder stores.
The following changes are part of this patch:
1) location URL format: previously cinder://<volume-id>, the format is
now changed to cinder://<store-name>/<volume-id>
2) Check for configured volume types: During initialization of service,
all volume types configured in the cinder store(s) will be checked if
they exist in deployment or not, the store with the wrong type will be
disabled. if cinder store is configured using traditional/old way
(i.e. using 'stores' config option) then this validation will not be performed.
3) Update legacy images: This change is partially depedent on the glance
changes[1] which trigger updating the URL for legacy images and the
cinder store checks the configured stores volume type against the existing
image-volume's type and update it respectively.

[1] https://review.opendev.org/#/c/742373/

Implements: blueprint multiple-cinder-backend-support
Change-Id: I0ba36d1095886fd1d96926af37a3a65a178c004f