Drop snapshot in use log from ERROR to WARN
This message is not actually an error inside the system for most of the
time.  The primary reason behind this is that it is possible that there
are VMs using this image and therefore it can't be deleted, the operator
has no actionable thing to do about this.

For those who are monitoring ERROR level messages and alerting on them,
this is noisy and not actionable, drop this error message to warning.

Change-Id: If5447c3f5b4f678943273c2bdb2e44cd25d5b226