Pass multipath config while creating connector object
Currently we only pass the configured multipath option when getting
the connector properties and do not pass it when creating the
connector object which in turn does the multipathing in os-brick.
So we are only passing multipathing option to cinder for exporting &
mapping the volume to paths but not while creating the connector object
and hence doing multipathing partially.
This patch passes that option and fixes the stated issue.

Related-Bug: #1863983

Co-Authored-By: Gorka Eguileor <geguileo@redhat.com>

Change-Id: I82979c47ec9ec1dd6cc864d3827ab219177251f8
(cherry picked from commit 04e5ead7c000211a4c10104ed2bb65c9df7681ae)