Add volume multiattach handling
We implemented cinder's new attachment API support with patch[1].
It was needed to add multiattach volume handling added with this
patch.
There is no special configuration change or user interference needed.
If a volume is of a multiattach type, then it will be handled as
a multiattach volume.

[1] https://review.opendev.org/c/openstack/glance_store/+/782200

Implements: blueprint attachment-api-and-multiattach-support
Closes-Bug: #1904546

Change-Id: Iffb825492a20fd877476acad05f817b399072f01