Add debug logs to cinder store
When debugging issues related to glance cinder store, there are
several calls to cinder and it becomes hard to determine which step
we are currently executing without going through cinder logs.
This patch adds some useful debug logs for the new attachment's code
to understand which stage of attachment we are on.

Change-Id: I491b7292a511c47c1d6148dab69ae04269e50c85