Remove redundant try except around volume create
The cinder volume create call was moved to the cinder_utils
module to keep the driver code cleaner with commit[1].
A decorator was also added to handle NotFound exception which
made the existing try/except block redundant.
Also the current try/except block has misleading messages:

1) "Invalid volume type configured"
2) "Failed to create image-volume due to invalid
   `cinder_volume_type` configured."

A volume create could fail because of many reasons and "invalid
volume type" is just one of them. These messages will confuse
the operators even if the configured volume type is correct.
This patch removes them since the NotFound exception is already
handled at[2] and the messages are misleading.

[1] 1178f113c4fc3f5f0938f3dcc608cea4a73adb7f
[2] https://github.com/openstack/glance_store/blob/c0c4969a4a6ce77c61f84fcb37b3efdac71d6418/glance_store/common/cinder_utils.py#L37

Change-Id: I39c9caec85f574fc717e60f11e5681860a20ed30