Do not always import boto3
Currently boto3 is not part of requirements but stevedore always tries
to import it and shows error in case boto3 is missing. This is not
a real error unless users actually enable s3 backends and can be quite
confusing.

This makes the driver code ignore ImportError and actually fail only
if users try to enable s3 backend without installing boto3.

Closes-Bug: #2007924
Change-Id: Ia94dd1d12a3d723f6263bdfb0966d416dfbae1af