Deprecate VMWare Datastore
The vmwareapi virt driver in nova was marked as experimental some
cycles ago due to lack of CI and maintainers[1]. This deprecates
the VMWare Datastore so that we can remove it when nova removes
the virt driver. The store driver is used along with the vmwareapi virt
driver and we don't expect any usage of this store driver with
the other virt drivers such as libvirt driver.

[1] 713ef7839670bd5d73af52c211a14932adc933ba

Change-Id: I57fc8cbd6308c4a82a638a2d1e997605736aedc4