Update maintainer of rbd driver and cinder driver
The rbd driver maintainer currently written in the doc is no longer
active in the community. Because the driver is used commonly in
deployments with Ceph, it should be kept maintained, so transfer its
maintainer-ship to the whole glance team.

Also, the previous cinder driver maintainer is no longer active and
Rajat agreed to replace him. Update the maintainer field for cinder
driver accordingly.

Change-Id: I2f0d9c743406d9f30594bb6dfd34809ab3c942df