Cinder PowerFlex: Fix issue when downloading image
When using Glance -> Cinder -> PowerFlex configuration, we fail when
trying to download an image. This regression was introduced with
2 changes:
1. Introduce wait mechanism for extended LUN[1]
2. Generalize the wait mechanism for all operations involving attach[2]

Originally the wait logic was introduced for create image operation but
later was generalized to get image operation as well with the
refactoring done in [2].

When extending the volume, the seek pointer is reset after the extend
so this wasn't an issue but when we try to download the image,
the file pointer stays to the end of file and we end up reading
nothing eventually failing the operation.

The reason for using seek and tell method is:
fstat doesn't work with block devices exported by the scaleio/powerflex
protocol as seen in the testing done.

$ os.fstat(device_file.fileno())

os.stat_result(st_mode=25008, st_ino=818, st_dev=5, st_nlink=1,
st_uid=42415, st_gid=6, st_size=0, st_atime=1731055942,
st_mtime=1731055942, st_ctime=1731055944)

See that st_size equals 0

Also the linux stat command returns 0 size:

Symbolic Link:
$ stat /dev/disk/by-id/emc-vol-01f948a4798c2e0f-e062e68a0000001d

  File: /dev/disk/by-id/emc-vol-01f948a4798c2e0f-e062e68a0000001d -> ../../scinib

  Size: 12              Blocks: 0          IO Block: 4096   symbolic link
...

Original File:
$ stat /dev/scinib

  File: /dev/scinib

  Size: 0               Blocks: 0          IO Block: 4096   block special file
...

To fix this download issue, we get the original position of file,
seek to the end to get the device size and eventually reset the
file pointer to the original position.

[1] https://review.opendev.org/c/openstack/glance_store/+/826290
[2] https://review.opendev.org/c/openstack/glance_store/+/843103

Closes-Bug: #2086759
Change-Id: I713b24a233e33d2908ac2706c9d5fea1c048810c
(cherry picked from commit 55cefedd169b84d713a8ff080d909483ea51c913)